* AB_ARRAY FIXED (working properly) * ScreenQuad does not requires tex coordinates for image rendering * Obj Loaded Fixed * TGA Loaded Added * Indirect Draw Buffer Added (in comments) * New models Added

+++ app/src/main/assets/Shaders/ab_array_init.frag
@@ -2,7 +2,7 @@
 
 precision highp float;
 
-layout(binding = 0, r32ui) uniform mediump writeonly uimage2D uniform_image_counter;
+layout(binding = 0, r32ui) uniform highp writeonly uimage2D     uniform_image_counter;
 
 void main()
 {
